Put gui generation into separate file
